"Changed the way a WhatsApp ID is constructed in connect function"
Explanation:
The code diff shows that the way a WhatsApp ID is constructed in the connect function has been changed. Previously, the ID was constructed by concatenating the "to" parameter with "@s.whatsapp.net". Now, the ID is simply set to the "to" parameter. It is unclear from the diff why this change was made, but the code itself has not been modified in any other way.