Update send message destination for chat and image messages
The code diff updates the destination address when sending chat and image messages using the `sendMessage` function. For chat messages, `to + "@s.whatsapp.net"` is used as the recipient. For image messages, `id` is used as the recipient in the `sendMessage` function call. Previously, the code used `to + "@s.whatsapp.net"` for image messages, which has been updated to `id`.