"Changed the way a WhatsApp ID is constructed in connect function"
Explanation:
The code diff shows that the way a WhatsApp ID is constructed in the `connect` function has been changed. Previously, the ID was constructed by concatenating the `to` parameter with "@s.whatsapp.net". Now, the `to` parameter itself is being used as the ID. This change may affect how the `sock.sendMessage` function is called in the following lines of code.